+++ examples/aws-native-ts-stepfunctions/world_function.js
@@ -1,0 +1,5 @@
+exports.handler = function(event, context, callback){
+    var response = event.response;
+    const updated = { "response": response + "World!" };
+    callback(null, updated);
+};