+++ provider/pkg/schema/convert.go
@@ -1,4 +1,16 @@
 // Copyright 2016-2021, Pulumi Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package schema
 
@@ -15,7 +27,7 @@ import (
 
 // SdkToCfn converts Pulumi-SDK-shaped state to CloudFormation-shaped payload. In particular, SDK properties
 // are lowerCamelCase, while CloudFormation is usually (but not always) PascalCase.
-func SdkToCfn(res *CloudAPIResource, types map[string]CloudAPIType, properties map[string]interface{}) map[string]interface{} {
+func SdkToCfn(res *CloudAPIResource, types map[string]CloudAPIType, properties map[string]interface{}) (map[string]interface{}, error) {
 	converter := sdkToCfnConverter{res, types}
 	return converter.sdkToCfn(properties)
 }
@@ -24,7 +36,7 @@ func SdkToCfn(res *CloudAPIResource, types map[string]CloudAPIType, properties m
 // mapped to corresponding patch terms, and SDK properties are translated to respective CFN names.
 func DiffToPatch(res *CloudAPIResource, types map[string]CloudAPIType, diff *resource.ObjectDiff) ([]jsonpatch.JsonPatchOperation, error) {
 	converter := sdkToCfnConverter{res, types}
-	return converter.diffToPatch(diff), nil
+	return converter.diffToPatch(diff)
 }
 
 type sdkToCfnConverter struct {
@@ -32,90 +44,124 @@ type sdkToCfnConverter struct {
 	types map[string]CloudAPIType
 }
 
-func (c *sdkToCfnConverter) sdkToCfn(properties map[string]interface{}) map[string]interface{} {
+func (c *sdkToCfnConverter) sdkToCfn(properties map[string]interface{}) (map[string]interface{}, error) {
 	result := map[string]interface{}{}
+	var err error
 	for k, prop := range c.spec.Inputs {
 		if v, ok := properties[k]; ok {
-			result[ToCfnName(k)] = c.sdkTypedValueToCfn(&prop.TypeSpec, v)
+			result[ToCfnName(k)], err = c.sdkTypedValueToCfn(&prop.TypeSpec, v)
+			if err != nil {
+				return nil, err
+			}
 		}
 	}
 	for k, attr := range c.spec.Outputs {
 		if v, ok := properties[k]; ok {
-			result[ToCfnName(k)] = c.sdkTypedValueToCfn(&attr.TypeSpec, v)
+			result[ToCfnName(k)], err = c.sdkTypedValueToCfn(&attr.TypeSpec, v)
+			if err != nil {
+				return nil, err
+			}
 		}
 	}
-	return result
+	return result, nil
 }
 
-func (c *sdkToCfnConverter) sdkTypedValueToCfn(spec *pschema.TypeSpec, v interface{}) interface{} {
+func (c *sdkToCfnConverter) sdkTypedValueToCfn(spec *pschema.TypeSpec, v interface{}) (interface{}, error) {
 	if spec.Ref != "" {
-		if spec.Ref == "pulumi.json#/Any" {
-			return v
+		switch spec.Ref {
+		case "pulumi.json#/Any":
+			return v, nil
+		case "pulumi.json#/Asset", "pulumi.json#/Archive":
+			switch t := v.(type) {
+			case *resource.Asset:
+				b, err := t.Bytes()
+				if err != nil {
+					return nil, err
+				}
+				return string(b), nil
+			}
 		}
 
 		typName := strings.TrimPrefix(spec.Ref, "#/types/")
 		return c.sdkObjectValueToCfn(typName, v)
 	}
 
+	var err error
 	switch spec.Type {
 	case "array":
 		array := v.([]interface{})
 		vs := make([]interface{}, len(array))
 		for i, item := range array {
-			vs[i] = c.sdkTypedValueToCfn(spec.Items, item)
+			vs[i], err = c.sdkTypedValueToCfn(spec.Items, item)
+			if err != nil {
+				return nil, err
+			}
 		}
-		return vs
+		return vs, nil
 	case "object":
 		sourceMap := v.(map[string]interface{})
 		vs := map[string]interface{}{}
 		for n, item := range sourceMap {
-			vs[n] = c.sdkTypedValueToCfn(spec.AdditionalProperties, item)
+			vs[n], err = c.sdkTypedValueToCfn(spec.AdditionalProperties, item)
+			if err != nil {
+				return nil, err
+			}
 		}
-		return vs
+		return vs, nil
 	default:
-		return v
+		return v, nil
 	}
 }
 
-func (c *sdkToCfnConverter) sdkObjectValueToCfn(typeName string, value interface{}) interface{} {
+func (c *sdkToCfnConverter) sdkObjectValueToCfn(typeName string, value interface{}) (interface{}, error) {
 	properties, ok := value.(map[string]interface{})
 	if !ok {
-		return value
+		return value, nil
 	}
 
 	spec := c.types[typeName]
 	result := map[string]interface{}{}
+	var err error
 	for k, prop := range spec.Properties {
 		if v, ok := properties[k]; ok {
-			result[ToCfnName(k)] = c.sdkTypedValueToCfn(&prop.TypeSpec, v)
+			result[ToCfnName(k)], err = c.sdkTypedValueToCfn(&prop.TypeSpec, v)
+			if err != nil {
+				return nil, err
+			}
 		}
 	}
-	return result
+	return result, nil
 }
 
-func (c *sdkToCfnConverter) diffToPatch(diff *resource.ObjectDiff) []jsonpatch.JsonPatchOperation {
+func (c *sdkToCfnConverter) diffToPatch(diff *resource.ObjectDiff) ([]jsonpatch.JsonPatchOperation, error) {
 	var ops []jsonpatch.JsonPatchOperation
 	for sdkName, prop := range c.spec.Inputs {
 		cfnName := ToCfnName(sdkName)
 		key := resource.PropertyKey(sdkName)
 		if v, ok := diff.Updates[key]; ok {
-			op := c.valueToPatch("replace", cfnName, prop, v.New)
+			op, err := c.valueToPatch("replace", cfnName, prop, v.New)
+			if err != nil {
+				return nil, err
+			}
 			ops = append(ops, op)
 		}
 		if v, ok := diff.Adds[key]; ok {
-			op := c.valueToPatch("add", cfnName, prop, v)
+			op, err := c.valueToPatch("add", cfnName, prop, v)
+			if err != nil {
+				return nil, err
+			}
 			ops = append(ops, op)
 		}
 		if _, ok := diff.Deletes[key]; ok {
-			op := jsonpatch.NewPatch("remove", "/" + cfnName, nil)
+			op := jsonpatch.NewPatch("remove", "/"+cfnName, nil)
 			ops = append(ops, op)
 		}
 	}
-	return ops
+	return ops, nil
 }
 
-func (c *sdkToCfnConverter) valueToPatch(opName, propName string, prop pschema.PropertySpec, value resource.PropertyValue) jsonpatch.JsonPatchOperation {
-	op := jsonpatch.NewPatch(opName, "/" + propName, nil)
+func (c *sdkToCfnConverter) valueToPatch(opName, propName string, prop pschema.PropertySpec, value resource.PropertyValue) (jsonpatch.JsonPatchOperation, error) {
+	op := jsonpatch.NewPatch(opName, "/"+propName, nil)
 	switch {
 	case value.IsNumber() && prop.Type == "integer":
 		i := int32(value.NumberValue())
@@ -128,12 +174,15 @@ func (c *sdkToCfnConverter) valueToPatch(opName, propName string, prop pschema.P
 		op.Value = value.StringValue()
 	default:
 		sdkObj := value.MapRepl(nil, nil)
-		cfnObj := c.sdkTypedValueToCfn(&prop.TypeSpec, sdkObj)
+		cfnObj, err := c.sdkTypedValueToCfn(&prop.TypeSpec, sdkObj)
+		if err != nil {
+			return jsonpatch.JsonPatchOperation{}, err
+		}
 		jsonBytes, err := json.Marshal(cfnObj)
 		contract.AssertNoError(err)
 		op.Value = string(jsonBytes)
 	}
-	return op
+	return op, nil
 }
 
 // CfnToSdk converts CloudFormation-shaped payload to Pulumi-SDK-shaped state. In particular, SDK properties