+++ sdk/nodejs/types/output.ts
@@ -16129,7 +16129,7 @@ export namespace lambda {
         /**
          * The source code of your Lambda function. If you include your function source inline with this parameter, AWS CloudFormation places it in a file named index and zips it to create a deployment package..
          */
-        zipFile?: string;
+        zipFile?: pulumi.asset.Asset | pulumi.asset.Archive;
     }
 
     /**